tests/wpt/css-tests/css-flexbox-1_dev/html/flexbox_computedstyle_flex-grow-0.htm

<!DOCTYPE html>
<html><head><title>flexbox | computed style | flex-grow: 0</title>
<link href="http://opera.com" rel="author" title="Opera Software">
<link href="http://www.w3.org/TR/css-flexbox-1/#propdef-flex-grow" rel="help">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta content="dom" name="flags">
<style>
body {
	color: red;
	flex-grow: 0;
}
</style>
</head><body><div id="log"></div>
<script>
test(function() {
    var body = document.body;

    assert_equals(getComputedStyle(body).getPropertyValue("flex-grow"),
	          "0");
});
</script>
</body></html>
